requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-cors-headers==3.13.0
exceptiongroup==1.1.0
filelock==3.9.0
frozenlist==1.3.3
gunicorn==20.1.0
huggingface-hub==0.12.1
idna==3.4
multidict==6.0.4
numpy==1.24.2
openai==0.27.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
trio==0.22.0
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
psycopg2-binary==2.9.5